[Feat] Soft Delete 구현
1. Question과 Answer Soft Delete 기능 개발을 위해 isDeleted 컬럼 추가
2. Soft Delete 시 Delete 요청 시간을 기록
3. Question 삭제 시, Answer도 삭제 되도록 더티체킹 사용하여 구현

+++ src/main/java/com/seonhansite/server/answer/AnswerService.java
@@ -33,12 +33,15 @@ public class AnswerService {
         if (answer.isPresent()) {
             return new AnswerResponse(answer.get());
         } else {
-            throw new DataNotFoundException("answer not found");
+            throw new DataNotFoundException("answer not found with id: "+ id);
         }
     }
 
     @Transactional
     public AnswerListResponse getList(Long id) {
+        Question question = this.questionRepository.findById(id)
+                .orElseThrow(() -> new DataNotFoundException("Question not found with id: " + id));
+
         List<AnswerResponse> al = this.answerRepository.findAllByQuestionId(id).stream().map(AnswerResponse::new).sorted(Comparator.comparing(AnswerResponse::getCreatedAt).reversed()).toList();
 
         int count = al.size();
@@ -85,6 +88,9 @@ public class AnswerService {
     public Integer deleteAnswer(Long id) {
         try {
             Answer answer = this.answerRepository.findById(id).orElseThrow(AnswerNotFoundException::new);
+            if (answer.getIsDeleted() != null) {
+                throw new AnswerNotFoundException();
+            }
             this.answerRepository.delete(answer);
             return 1;
         } catch (AnswerNotFoundException e) {
